Misc Reorg (#68)
* reorganize geometry code

* add shortcuts to common classes

* move quaternion yaw back
