+++ python-sdk/nuscenes/eval/detection/__init__.py
@@ -1,0 +1,1 @@
+from .main import NuScenesEval
