+++ python-sdk/nuscenes/eval/detection/render.py
@@ -3,7 +3,7 @@ from typing import Dict
 import numpy as np
 from matplotlib import pyplot as plt
 from nuscenes.eval.detection.utils import boxes_to_sensor
-from nuscenes.nuscenes import NuScenes
+from nuscenes import NuScenes
 from nuscenes.utils.data_classes import LidarPointCloud
 from nuscenes.utils.geometry_utils import view_points
