+++ requirements.txt
@@ -3,7 +3,7 @@ matplotlib==2.2.3
 numpy==1.14.5
 opencv-python==3.4.2.17
 Pillow==5.2.0
-pyquaternion==0.9.2
+pyquaternion==0.9.5
 scikit-learn==0.19.2
 tqdm==4.25.0
 scipy==1.1.0
